% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_longate.R
\name{df_longate}
\alias{df_longate}
\title{Pivot wider treatment practice input data}
\usage{
df_longate(df)
}
\arguments{
\item{df}{A data.frame, treatment practice inputs}
}
\value{
A long format data.frame
}
\description{
Pivot wider treatment practice input data
}
