% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLS_Adherence_100.R
\name{FLS_Adherence_100}
\alias{FLS_Adherence_100}
\title{Preparing FLS_Adherence_100 model inputs}
\usage{
FLS_Adherence_100(df_sheet)
}
\arguments{
\item{df_sheet}{A data.frame, googlesheet equivalent of base case}
}
\value{
A data.frame
}
\description{
Preparing FLS_Adherence_100 model inputs
}
