% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLS_Injectables_onlymax_reduction.R
\name{FLS_Injectables_onlymax_reduction}
\alias{FLS_Injectables_onlymax_reduction}
\title{Preparing FLS_Injectables_onlymax_reduction model inputs}
\usage{
FLS_Injectables_onlymax_reduction(df_sheet)
}
\arguments{
\item{df_sheet}{A data.frame, googlesheet equivalent of base case}
}
\value{
A data.frame
}
\description{
Preparing FLS_Injectables_onlymax_reduction model inputs
}
