% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hips_and_spine_Only.R
\name{Hips_and_spine_Only}
\alias{Hips_and_spine_Only}
\title{Preparing FLS Hips_and_spine_Only model inputs}
\usage{
Hips_and_spine_Only(df_sheet)
}
\arguments{
\item{df_sheet}{A data.frame, googlesheet equivalent of base case}
}
\value{
A data.frame
}
\description{
Preparing FLS Hips_and_spine_Only model inputs
}
