% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hips_only.R
\name{Hips_only}
\alias{Hips_only}
\title{Preparing FLS Hips_only model inputs}
\usage{
Hips_only(df_sheet)
}
\arguments{
\item{df_sheet}{A data.frame, googlesheet equivalent of base case}
}
\value{
A data.frame
}
\description{
Preparing FLS Hips_only model inputs
}
