% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_numOut.R
\name{nice_numOut}
\alias{nice_numOut}
\title{Output long numeric values nicely e.g 10,000 instead of 10000}
\usage{
nice_numOut(dframe)
}
\arguments{
\item{dframe}{A data.frame or tibble}
}
\value{
A data frame
}
\description{
Output long numeric values nicely e.g 10,000 instead of 10000
}
