% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmacological_vals_sum_check.R
\name{pharmacological_vals.sum.check}
\alias{pharmacological_vals.sum.check}
\title{Sum 1 check for select pharmacological values}
\usage{
pharmacological_vals.sum.check(
  df_sheet,
  base_vars = NULL,
  gender = NULL,
  site = NULL,
  arm = NULL
)
}
\arguments{
\item{df_sheet}{Scenario value inputs}

\item{base_vars}{A vector of pharmacological values}

\item{gender}{Either \code{male} or \code{female}}

\item{site}{One of \code{hip}, \code{spine} or \code{other}}

\item{arm}{One of \code{fls} or \code{no.fls}}
}
\value{
A vector sum 1's
}
\description{
Sum 1 check for select pharmacological values
}
