% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refrac5_10year.R
\name{refrac5_10year}
\alias{refrac5_10year}
\title{Contingency tables for 5 & 10 year re-fracture risk model inputs}
\usage{
refrac5_10year(dframe = NULL, index = NULL)
}
\arguments{
\item{dframe}{A data.frame, model specific inputs from googlesheet}

\item{index}{Fracture site, one of \code{hip}, \code{spine}, \code{other}}
}
\value{
A data.frame
}
\description{
Contingency tables for 5 & 10 year re-fracture risk model inputs
}
