% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_current_practice.R
\name{replace_current_practice}
\alias{replace_current_practice}
\title{Replace CP table values in FLS adherence and perfect FLS}
\usage{
replace_current_practice(table_read = NULL, scenario_type = scenario_data.path)
}
\arguments{
\item{table_read}{FLS model output table i.e \code{summary.discounted.qol.t60.csv}}

\item{scenario_type}{FLS scenarion run}
}
\value{
A data.frame
}
\description{
Replace CP table values in FLS adherence and perfect FLS
}
