% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staff_mins_time.R
\name{staff_mins_time}
\alias{staff_mins_time}
\title{Tabulating staff minutes for all fracture sites}
\usage{
staff_mins_time(fract_site = "hip")
}
\arguments{
\item{fract_site}{A character value for fracture site. \code{hip}, \code{spine} or \code{other}}
}
\value{
A data.frame
}
\description{
Tabulating staff minutes for all fracture sites
}
